Build fixes, now compiles with no warnings (with gcc 5 default warning level, and -Wno-unused-result)

+++ src/mainfuncs.h
@@ -24,7 +24,7 @@ struct indexlist {
 };
 
 struct menu_struct {
-	int  shiftcount;
+	size_t shiftcount;
 	int  able;
 	char display[MAINLINE];
 	char keys[MAINLINE];
